// types/ourwork/specificpageDTO.ts
export interface SpecificPageDTO {
  sectionOne: Section;
  sectionTwo: Section;
  sectionThree: Section;
  sectionFour: Section;
  sectionFive: Section;
  sectionSix: Section;
  sectionSeven: Section;
  sectionEight: Section;
  sectionNine: Section;
  sectionTen: Section;
}

export interface Section {
  id: string;
  position: number;
}